dom/tests/mochitest/bugs/test_bug817476.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=817476
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 817476</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=817476">Mozilla Bug 817476</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 817476 **/
function getNewWindow() {
  var ifr = document.createElement("iframe");
  $("content").appendChild(ifr);
  return ifr.contentWindow;
}

// Test getting .screen before .Screen
var win = getNewWindow();
is(Object.getPrototypeOf(win.screen), win.Screen.prototype,
   "protos must match (1)");
is(win.Screen.prototype.toString(), "[object ScreenPrototype]",
   "proto must be WebIDL (1)");

// Test getting Screen before .screen
var win = getNewWindow();
is(win.Screen.prototype, Object.getPrototypeOf(win.screen),
   "protos must match (2)");
is(win.Screen.prototype.toString(), "[object ScreenPrototype]",
   "proto must be WebIDL (2)");

</script>
</pre>
</body>
</html>
